Styling revamp; relocate clients/edit
