+++ app/views/users/events/_event.html.erb
@@ -1,13 +1,17 @@
 <% if event.message_id? %>
-  <p class="incoming-msg-from">Client <%= event.created_at.strftime("%m/%d/%Y %l:%M %p") %></p><div class="clear"></div>
   <div class="text-bubble incoming">
     <%= event.content %>
   </div>
-  <div class="msg-sep"></div>
+  <p class="incoming-msg-from">
+    <%= event.created_at.strftime("%m/%d/%Y %l:%M %p") %>
+  </p>
 <% else %>
-  <p class="incoming-msg-to">Tony P. <span class="provider-name">Provider</span> <%= event.created_at.strftime("%m/%d/%Y %l:%M %p") %></p><div class="clear"></div>
-  <div class="text-bubble">
+  <div class="text-bubble outgoing">
     <%= event.content %>
   </div>
-  <div class="msg-sep"></div>
+  <p class="incoming-msg-to">
+    Tony P.
+    <span class="provider-name">Provider</span>
+    <%= event.created_at.strftime("%m/%d/%Y %l:%M %p") %>
+  </p>
 <% end %>
