+++ app/views/users/interactions/_referral.html.erb
@@ -5,11 +5,14 @@
       options_for_select(["Emergency Shelter", "Food Programs & Pantries", "Utility Assistance"]),
       id: "service-type-search",
       class: "selectize sans-body",
-      prompt: "Select A Service Type"  %>
+      prompt: "Select A Service Type" %>
   </div>
 <% else %>
   <div class="referral-send-container">
-    <p>Refer Client To Services:</p>
-    <button class="send-referral button" data-referral-code="<%= @interaction.referral.code %>"><%= "Send Code #{@interaction.referral.code} to Client" %></button>
+    <h1>Referal code: <%= @interaction.referral.code %></h1>
+
+    <button class="send-referral c-btn" data-referral-code="<%= @interaction.referral.code %>">
+      <%= "Send to Client" %>
+    </button>
   </div>
 <% end %>
