--- app/views/users/interactions/clients/edit.html.erb
@@ -1,64 +1,0 @@
-<div class="profile-container">
-  <%= form_for [:users, @interaction, @client] do |f| %>
-    <div class="profile-top">
-      <div>
-        <%= f.label :first_name %>
-        <%= f.text_field :first_name %>
-      </div>
-      <div>
-        <%= f.label :last_name %>
-        <%= f.text_field :last_name %>
-      </div>
-      <div>
-        <%= f.label :birthdate %>
-        <%= f.date_select :birthdate, {order: [:month, :day, :year], prompt: { day: 'Day', month: 'Month', year: 'Year' }, start_year: Date.today.year - 18, end_year: Date.today.year - 100}, {required: true} %>
-      </div>
-      <div>
-        <%= f.label :last_address %>
-        <%= f.text_field :last_address %>
-      </div>
-      <div>
-        <%= f.label :phone %>
-        <%= f.phone_field :phone %>
-      </div>
-      <div>
-        <%= f.label :gender %>
-        <%= f.text_field :gender %>
-      </div>
-      <div>
-        <%= f.label :ssn %>
-        <%= f.text_field :ssn %>
-      </div>
-      <div>
-        <%= f.label :ethnicity %>
-        <%= f.text_field :ethnicity %>
-      </div>
-      <div>
-        <%= f.label :education %>
-        <%= f.text_field :education %>
-      </div>
-      <div>
-        <%= f.label :employed %>
-        <%= f.check_box :employed %>
-      </div>
-      <div>
-        <%= f.label :veteran %>
-        <%= f.check_box :veteran %>
-      </div>
-      <div>
-        <%= f.label :domestic_assault_victim %>
-        <%= f.check_box :domestic_assault_victim %>
-      </div>
-      <div>
-        <%= f.label :mental_health %>
-        <%= f.check_box :mental_health %>
-      </div>
-      <div>
-        <%= f.label :kid_count %>
-        <%= f.number_field :kid_count %>
-      </div>
-
-      <%= f.submit "Submit" %>
-    </div>
-  <% end %>
-</div>
