+++ app/views/users/interactions/index.html.erb
@@ -1,5 +1,5 @@
 <div class="container-center">
-  <%= link_to "New Interaction", new_users_interaction_url %>
+  <%= link_to "New Interaction", new_users_interaction_url, class: "c-btn new-interaction-btn" %>
 
   <p class="interaction-section-header">Active</p>
   <ul class="interactions-container">
