+++ app/views/users/interactions/show.html.erb
@@ -1,7 +1,7 @@
 <script type="text/javascript">
   $(function() {
     $("form.new_event input[type='submit']").click(function(e){
-      var replyBody = $("textarea#event_content").val();
+      var replyBody = $("input#event_content").val();
       if (0 === replyBody.length) {
         e.preventDefault();
         alert("You must enter a reply.")
@@ -10,7 +10,7 @@
 
     $(".referral-send-container button.send-referral").click(function() {
       var code = $(this).data("referral-code");
-      $("textarea#event_content").val(code);
+      $("input#event_content").val("Your code is: "+code);
     })
   })
 </script>
@@ -22,8 +22,8 @@
       <%= f.hidden_field :user_id %>
       <%= f.hidden_field :source_type, value: 'sms' %>
 
-      <%= f.text_area :content %>
-      <%= f.submit "Reply", class: "button" %>
+      <%= f.text_field :content %>
+      <%= f.submit (@interaction.type == "sms" ? "Reply" : "Note"), class: 'c-btn' %>
     <% end %>
   </div>
