+++ app/views/users/referrals/new.html.erb
@@ -5,5 +5,5 @@
   <%= f.hidden_field :interaction_id %>
   <%= f.hidden_field :service_id %>
 
-  <%= f.submit "Create" %>
+  <%= f.submit "Create", class: "c-btn" %>
 <% end %>
